requirements.txt
boto3
titus2
